# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Laravel API'
  description: 'Laravel APIの概要'
  version: 1.0.0
servers:
  - url: http://localhost/api
paths:
  '/user':
    get:
      tags:
        - User
      summary: ユーザー情報を取得する
      description: '現在のユーザーの情報を取得 (GET)'
      operationId: getUser
      security:
        - BearerAuth: []  # Authorizationヘッダーでトークンを送信する
      responses:
        '200':
          description: ユーザー情報が正常に取得されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthenticated.

    put:
      tags:
        - User
      summary: ユーザー情報を更新する
      description: '現在のユーザーの情報を更新 (PUT)'
      operationId: updateUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - password
                - password_confirmation
              properties:
                name:
                  type: string
                  description: 更新後のユーザー名
                  example: New Name
                password:
                  type: string
                  description: 更新後のパスワード
                  example: password123
                password_confirmation:
                  type: string
                  description: パスワード確認用
                  example: password123
      responses:
        '200':
          description: ユーザー情報が正常に更新されました
        '400':
          description: バリデーションエラー
        '401':
          description: 認証されていません。
    delete:
      tags:
        - User
      summary: ユーザーを削除する
      description: '現在のユーザーを削除 (DELETE)'
      operationId: deleteUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ユーザーが正常に削除されました
        '401':
          description: 認証されていません。
  '/register/step1':
    post:
      tags:
        - User
      summary: ユーザー仮登録
      description: 'ユーザー登録の仮登録ステップ (POST)'
      operationId: registerStep1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
                  description: 登録するメールアドレス
                  example: johndoe@example.com
              type: object
      responses:
        '200':
          description: 仮登録が完了しました。メールをご確認ください。
        '400':
          description: 無効なメールアドレスまたは既に登録済み
  '/register/step2':
    post:
      tags:
        - User
      summary: ユーザー本登録
      description: 'ユーザー登録の本登録ステップ (POST)'
      operationId: registerStep2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - token
                - name
                - password
              properties:
                token:
                  type: string
                  description: 仮登録時に送信されたトークン
                  example: abcdef123456
                name:
                  type: string
                  description: ユーザー名
                  example: John Doe
                password:
                  type: string
                  description: パスワード
                  example: password123
              type: object
      responses:
        '201':
          description: 本登録が完了しました。
        '400':
          description: 入力が無効です。
        '401':
          description: トークンが無効です。
  '/login':
    post:
      tags:
        - User
      summary: ユーザーログイン
      description: '既存のユーザーがログイン (POST)'
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  description: メールアドレス
                  example: johndoe@example.com
                password:
                  type: string
                  description: パスワード
                  example: password123
              type: object
      responses:
        '200':
          description: ログインが成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: 認証トークン
                    example: 1|XyZ1234ExampleToken
                  token:
                    type: string
                    description: ユーザー名
                    example: Name
        '400':
          description: 入力が無効です。
        '401':
          description: パスワードが違います。
  '/logout':
    post:
      tags:
        - User
      summary: ユーザーをログアウトする
      description: '現在のユーザーをログアウトします'
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: ログアウトが成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: ログアウトが成功しました
        '401':
          description: 認証されていません
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ユーザーID
          example: 
        name:
          type: string
          description: ユーザー名
          example: John Doe
        email:
          type: string
          description: メールアドレス
          example: johndoe@example.com
        created_at:
          type: string
          format: date-time
          example: '2024-12-01T00:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-12-02T00:00:00Z'
...